+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * `Decoder::new_unbuffered` and `Encoder::new_unbuffered`
 ### Changed
 * `Decoder::new` and `Encoder::new` are now unbuffered
+* Replaced lzw crate usage with new, faster implmentatiton
+
 ## [0.9.0] - 2020-06-04
 ### Added
 * Result type for gift crate